# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "GamiBook API",
        "description": "ESMAPP API - GamiBook é uma aplicação para a editora LIDEL que incentiva a aprendizagem através de exercícios gamificados. O token não expira já que a aplicação é destinada para o uso mobile.",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
        "parameters": {
...
            "idUser": {
                "name": "idUser",
                "in": "path",
                "description": "Identificador do utilizador",
                "required": true,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            },
...
            "bookCategory": {
                "name": "category",
                "in": "query",
                "description": "Identificador da categoria",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            },
            "idBook": {
                "name": "idBook",
                "in": "path",
                "description": "Identificador do livro",
                "required": true,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            },
...
        },
...
    },
...
    "paths": {
...
        "/users": {
            "get": {
                "summary": "Lista de utilizadores",
                "description": "Devolve uma lista de utilizadores da aplicação",
                "operationId": "findAllUsers",
                "tags": [
                    "Utilizador",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/users/200/findAllUsers"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    }
                }
            },
            "post": {
                "summary": "Criar utilizadores administradores",
                "description": "Adiciona um utilizador à base de dados da aplicação",
                "operationId": "createAdminUser",
                "tags": [
                    "Utilizador",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/users/register_admin_user"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/users/201/createAdminUser"
                    },
                    "400": {
                        "$ref": "#/components/responses/users/400/createAdminUser"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    }
                }
            }
        },
...
        "/users/{idUser}": {
            "get": {
                "summary": "Procurar um utilizador",
                "description": "Devolve a informação de um utilizador pelo id",
                "operationId": "findOneUser",
                "tags": [
                    "Utilizador",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idUser"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/users/200/findOneUser"
                    },
                    "400": {
                        "$ref": "#/components/responses/general/400/idUser"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/idUser"
                    }
                }
            }
        },
...
        "/books": {
            "get": {
                "summary": "Lista de livros",
                "description": "Devolve uma lista de livros",
                "operationId": "findAllBooks",
                "tags": [
                    "Livro",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/bookCategory"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/books/200/findAllBooks"
                    },
                    "400": {
                        "$ref": "#/components/responses/general/400/idCategory"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    }
                }
            },
            "post": {
                "summary": "Criar livros",
                "description": "Adiciona um livro à base de dados da aplicação",
                "operationId": "createBook",
                "tags": [
                    "Livro",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/books/createBook"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/books/201/createBook"
                    },
                    "400": {
                        "$ref": "#/components/responses/books/400/createBook"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "406": {
                        "$ref": "#/components/responses/books/406/name_code"
                    }
                }
            }
        },
        "/books/{idBook}": {
            "get": {
                "summary": "Procurar um livro",
                "description": "Devolve a informação de um livro pelo id",
                "operationId": "findOneBook",
                "tags": [
                    "Livro",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idBook"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/books/200/findOneBook"
                    },
                    "400": {
                        "$ref": "#/components/responses/general/400/idBook"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/idBook"
                    }
                }
            },
            "patch": {
                "summary": "Atualizar o livro",
                "description": "Atualiza alguns campos do livro",
                "operationId": "updateBook",
                "tags": [
                    "Livro",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idBook"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/books/updateBook"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/books/200/updateBook"
                    },
                    "400": {
                        "$ref": "#/components/responses/books/400/updateBook"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/idBook"
                    }
                }
            },
            "delete": {
                "summary": "Apagar livros",
                "description": "Apaga um livro pelo id",
                "operationId": "deleteBook",
                "tags": [
                    "Livro",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idBook"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/books/200/deleteBook"
                    },
                    "400": {
                        "$ref": "#/components/responses/general/400/idBook"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/idBook"
                    }
                }
            }
        },
...
    }
}